works with Cygwin under Python 2.1 now, although there is still a problem:
The only way I know of to generate the Python import library for Cygwin is to
build a complete Python under cygwin. The result seems to use the cygwin
conventions for path parsing, so even if you build it you can't run it through
the standard NT run actions. Perhaps try MINGW32.

SVN-Revision: 10488