Make composing generators fail immediately if given no name. This, essentially, makes composing generators fail if they are run deeper during transformation search. Looks like a new incarnation of 'allow-composing'.
The motivation is to fight with cases where trasformation
like CPP -> OBJ -> STATIC_LIB -> RSP -> EXE was tried. The composing
OBJ -> STATIC_LIB generator was totally unprepared to be invoked at such
depth, and errored out. The new behaviour cause this transformation to be
rejected, but the search continues.

* new/generators.jam
  (generator.run-really): New rule, with most of previous 'run' content.
  (generator.run): Check arguments only and call 'run-really'. Immediately
  return if the generator is composing and no name is specified.

SVN-Revision: 18216