Preliminary implementation of the Launcher, Child and Status concepts. This is still *very* incomplete and highly subject to change but some functionality starts to work.  I'm commiting this to be able to continue development incrementally in a controlled environment and to avoid losing what already works.  I've been working on this code offline for several days.
Note that one of the unit tests does not pass.  This is because the user
has no way to close a communication pipe with a chile process, effectively
being unable to send an EOF to, e.g., stdin.  In order to fix this I have
to change multiple things that I'd rather do with the current code in a
safe place.

SVN-Revision: 5267