Get rid of the id_t and desc_t types.
The former is now into the basic_process class and is named handle_type,
following the same changes done in pipe and systembuf (I'm going to
document the rationale behind handles soon afterwards).

The latter simply disappears for now (becoming a regular int) to let me
remove the types.hpp header.  The descriptor thing needs a complete
overhaul after seeing some "limitations" in how Win32 communicates with
child processes, so this will be reviewed in depth later on.

SVN-Revision: 5352