Change the approach to represent the environment.  Previously, the code kept a list of "differences" between the current environment and the one to be configured for the child process.  This is appropriate for POSIX systems because such differences can be easily applied after the fork(2).
However, this makes implementation on a Win32 system extremely difficult,
if not impossible.  It also makes the procedure of cleaning the
environment for the child process quite complex.

To solve this, represent the environment as the list of all variables
that it will contain; therefore, the constructor initializes the class
with the current environment contents.  This makes it quite easy to
add variables to convert to the "char** envp" format and the strings
used by Win32.

Note that this breaks the current launcher; will fix it later.

SVN-Revision: 5385