factor out the search through Wow6432node in the registry of software
Make path-to-native work on zero or more paths, which simplifies lots
of code.

Factor native path translation out of compute-default-paths and into its caller.

Remove the path-to-native translation on the prefix from
candidate-interpreters, as the passed prefix has to be native anyway.

Fix the capturing of the exec_prefix for use in setting dll-path for
windows targets.

SVN-Revision: 37255