Revert commit 1367.
Go back to using getters in the status and posix_status classes instead of
public members.  Will solve some problems I was having when assigning such
objects and overall is a nicer design.

As a later step, it should be considered whether using boost::optional makes
sense as a return value for the functions, instead of having getters to query
the validity of others' result values.

SVN-Revision: 6803