Two unit tests suggested by Janek Kozicki.
- First, we test what happens if we load two times a library with the same name but with different implementation.
We load a first version named libHelloWorldCache.extension, we use it, and then we replace this file
with another library (but leaving the same name, libHelloWorldCache.extension) and check if the returned values
are the same (cached). In my tests on Linux it worked correctly (giving two diferent results).

- Second, we test what happens if we remove a shared object before unloading the plugin. It works well on Linux
too.

SVN-Revision: 7047