Added diagnostic message to output.c to inform users when a process has timed out and been killed.
Because timed out processes now emit a diagnostic, I
had to update option_l.jam so we wouldn't break test.sh
when it runs.

Minor cleanup to execunix.c to remove unneeded code
and to set the process exit status as returned from
waitpid.  The exit status is used to identify timed
out processes so we can emit a diagnostic to the user.

SVN-Revision: 39423