Implement PRECIOUS builtin.
Note that existing code already tried to assume that the 'together'
flag means precious, but as described at:
http://public.perforce.com:8080/@md=d&cd=//public/jam/src/&cdf=//public/jam/src/make1.c&c=klD@/1494?ac=10
this was not really wrong.

Now, PRECIOUS makes the target never removed. I plan to extend it so that
user can control what exit codes cause the target to be removed.

SVN-Revision: 60157