Renaming kolmogorov_max_flow to boykov_kolmogorov_max_flow.
At the request of the authors of the published algorithm, the header
and all associated functions, data types, tests, examples, and docs
should be renamed to boykov_kolmogorov. Branched all of the necessary
documents and renamed all such functions and data types.

Added deprecation warnings to the kolmogorov_max_flow.hpp and to the
kolmogorov_max_flow.html.

SVN-Revision: 63189