Generate home page in advance.
I've tweaked the php so it can be run from the command line, replacing
calls to `virtual` with SSI includes. Use that to generate the home
page. Might do this elsewhere, but would need to automate it. Just doing
the home page for now as it is rarely updated, and must be one of the
most popular pages. The 'recent announcements' at the bottom right are
still dynamically generated - could possibly run php from their cron job
to make them static.

SVN-Revision: 64729