Merge optimization for stacking macros.
It turns out the spirit 2 version didn't have the speed problem because the
spirit 2 macro copy constructs by reference. But this also meant that scoping
wasn't working properly. So this fixes that, but probably makes this slower.
Especially since gcc can't cope with a symbol table stored directly in a
container (because it tries to use `&` to take its address).

SVN-Revision: 65203